(ROX-16687) Remove NACK handling in Compliance as no longer needed
