+++ tools/local-compliance/main.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/stackrox/rox/compliance/collection/compliance"
-	"github.com/stackrox/rox/generated/internalapi/sensor"
 	"github.com/stackrox/rox/pkg/env"
 	"github.com/stackrox/rox/pkg/logging"
 	"github.com/stackrox/rox/pkg/retry"
@@ -24,11 +23,10 @@ func main() {
 	}
 	log.Infof("Generation inverval: %v", scanner.generationInterval.String())
 
-	srh := &dummySensorReplyHandlerImpl{}
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	nsr := retry.NewUnconfirmedMessageHandler(ctx, 5*time.Second)
-	c := compliance.NewComplianceApp(np, scanner, srh, nsr)
+	c := compliance.NewComplianceApp(np, scanner, nsr, nsr)
 	c.Start()
 }
 
@@ -36,16 +34,4 @@ type dummyNodeNameProvider struct{}
 
 func (dnp *dummyNodeNameProvider) GetNodeName() string {
 	return "local-compliance"
-}
-
-type dummySensorReplyHandlerImpl struct{}
-
-// HandleACK handles ACK message from Sensor
-func (s *dummySensorReplyHandlerImpl) HandleACK(_ context.Context, _ sensor.ComplianceService_CommunicateClient) {
-	log.Debugf("Received ACK from Sensor.")
-}
-
-// HandleNACK handles NACK message from Sensor
-func (s *dummySensorReplyHandlerImpl) HandleNACK(_ context.Context, _ sensor.ComplianceService_CommunicateClient) {
-	log.Infof("Received NACK from Sensor.")
 }
